requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
Django==4.0.7
django-celery-beat==2.3.0
django-timezone-field==5.0
djangorestframework==3.14.0
flower==1.2.0
humanize==4.4.0
kombu==5.2.4
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psycopg2==2.9.3
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.3
tornado==6.2
tzdata==2022.4
vine==5.0.0
wcwidth==0.2.5
